rename /feed to /activity, remove groups tab, add /feed redirect
