+++ frontend/app/(general)/activity/loading.tsx
